Add GPT-5.5 support via Azure Foundry and fix Next.js 15 async issues
- Added GPT-5.5 to the ModelSelectionForm UI.
- Created lib/utils/model.ts and refactored getModel to support Azure Foundry.
- Prioritized Azure in the model fallback chain.
- Fixed Next.js 15 async cookies issue in lib/auth/get-current-user.ts.
- Updated all agent import paths for getModel.
- Updated .env.local.example with Azure configuration variables.
