+++ lib/agents/query-suggestor.tsx
@@ -3,7 +3,7 @@ import { CoreMessage, LanguageModel, streamObject } from 'ai'
 import { PartialRelated, relatedSchema } from '@/lib/schema/related'
 import { Section } from '@/components/section'
 import SearchRelated from '@/components/search-related'
-import { getModel } from '../utils'
+import { getModel } from '../utils/model'
 
 interface CacheEntry {
   data: PartialRelated;
